backend-trading/requirements.txt
fastapi
uvicorn
python-telegram-bot
pandas
ta
requests
twelvedata